refactor: ChatRoomDto 패키지 이동
